Add company login call
