+++ sdk/build.gradle
@@ -24,17 +24,20 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    // Services
+    // SERVICES
     implementation 'com.squareup.retrofit2:retrofit:2.5.0'
     implementation 'com.squareup.retrofit2:converter-scalars:2.1.0'
     implementation 'com.google.code.gson:gson:2.8.2'
     implementation 'com.crashlytics.sdk.android:crashlytics:2.9.3', { transitive = true }
 
-    implementation 'com.android.support:appcompat-v7:28.0.0'
+    // TEST
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+
+    // GENERAL
+    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.3.11'
 }
 
 repositories {
