+++ sdk/src/main/java/com/ingresse/sdk/IngresseService.kt
@@ -2,8 +2,10 @@ package com.ingresse.sdk
 
 import com.ingresse.sdk.services.AuthService
 import com.ingresse.sdk.services.EntranceService
+import com.ingresse.sdk.services.UserService
 
 class IngresseService(client: IngresseClient) {
     var entrance: EntranceService = EntranceService(client)
     var auth: AuthService = AuthService(client)
+    var user: UserService = UserService(client)
 }
