+++ sdk/src/main/java/com/ingresse/sdk/model/response/CompanyApplicationJSON.kt
@@ -1,0 +1,7 @@
+package com.ingresse.sdk.model.response
+
+data class CompanyApplicationJSON(
+        val id: Int = 0,
+        val name: String = "",
+        val publicKey: String = ""
+)
