+++ sdk/src/main/java/com/ingresse/sdk/model/response/CompanyJSON.kt
@@ -1,0 +1,7 @@
+package com.ingresse.sdk.model.response
+
+data class CompanyJSON(
+        val id: Long = 0,
+        val name: String = "",
+        val logo: CompanyLogoJSON?
+)
