+++ sdk/src/main/java/com/ingresse/sdk/model/response/CompanyLoginJSON.kt
@@ -1,7 +1,7 @@
 package com.ingresse.sdk.model.response
 
-class LoginJSON {
+class CompanyLoginJSON {
     val status: Boolean = false
     val message: String? = null
-    val data: UserJSON? = null
+    val data: Array<CompanyDataJSON>? = null
 }
