+++ sdk/src/main/java/com/ingresse/sdk/model/response/CompanyLogoJSON.kt
@@ -1,0 +1,7 @@
+package com.ingresse.sdk.model.response
+
+data class CompanyLogoJSON(
+        val small: String = "",
+        val medium: String = "",
+        val large: String = ""
+)
