+++ sdk/src/main/java/com/ingresse/sdk/request/Auth.kt
@@ -5,20 +5,14 @@ import retrofit2.http.*
 
 interface Auth {
     /**
-     * Execute login for user
+     * Execute Company login for user
      *
      * @param email - User's email
-     * @param password = User's password
+     * @param password - User's password
      */
     @FormUrlEncoded
-    @POST("/login")
-    fun loginWithEmail(@Query("apikey") apikey: String,
+    @POST("/company-login")
+    fun companyLoginWithEmail(@Query("apikey") apikey: String,
                        @Field("email") email: String,
                        @Field("password") password: String) : Call<String>
-
-    @GET("/user/{userId}")
-    fun getUserData(@Path("userId") userId: Int,
-                    @Query("apikey") apikey: String,
-                    @Query("usertoken") userToken: String,
-                    @Query("fields") fields: String) : Call<String>
 }
