+++ sdk/src/main/java/com/ingresse/sdk/services/EntranceService.kt
@@ -71,4 +71,3 @@ class EntranceService(private val client: IngresseClient) {
         mGuestListCall?.enqueue(RetrofitCallback(type, callback))
     }
 }
-
